+++ Src/Private/Get-AbrVb365ObjectRepository.ps1
@@ -5,7 +5,7 @@ function Get-AbrVB365ObjectRepository {
     .DESCRIPTION
         Documents the configuration of Veeam VB365 in Word/HTML/Text formats using PScribo.
     .NOTES
-        Version:        0.3.2
+        Version:        0.3.4
         Author:         Jonathan Colon
         Twitter:        @jcolonfzenpr
         Github:         rebelinux
@@ -24,7 +24,7 @@ function Get-AbrVB365ObjectRepository {
 
     process {
         try {
-            $script:ObjectRepositories = Get-VBOObjectStorageRepository | Sort-Object -Property Name
+            $script:ObjectRepositories = Get-VBOObjectStorageRepository -WarningAction SilentlyContinue | Sort-Object -Property Name
             if (($InfoLevel.Infrastructure.Repository -gt 0) -and ($ObjectRepositories)) {
                 Write-PScriboMessage "Collecting Veeam VB365 Object Repository."
                 Section -Style Heading2 'Object Repositories' {
@@ -71,6 +71,7 @@ function Get-AbrVB365ObjectRepository {
                                     Paragraph {
                                         Text "Best Practice:" -Bold
                                         Text "Veeam recommend to implement Immutability where it is supported. It,s done for increased security: immutability protects your data from loss as a result of attacks, malware activity or any other injurious actions."
+                                        Text "https://bp.veeam.com/vb365/guide/design/hardening/Repo_specifics.html" -Color Blue
                                     }
                                     BlankLine
                                 }
@@ -98,6 +99,7 @@ function Get-AbrVB365ObjectRepository {
                             Paragraph {
                                 Text "Best Practice:" -Bold
                                 Text "Veeam recommend to implement Immutability where it is supported. It,s done for increased security: immutability protects your data from loss as a result of attacks, malware activity or any other injurious actions."
+                                Text "https://bp.veeam.com/vb365/guide/design/hardening/Repo_specifics.html" -Color Blue
                             }
                             BlankLine
                         }